Type cursor as u32

+++ nexus-webapi/src/routes/v0/events.rs
@@ -14,7 +14,7 @@ use utoipa::OpenApi;
 #[derive(Debug, Clone, Serialize, Deserialize, ToSchema, utoipa::ToResponse)]
 #[schema(as = String)]
 pub struct EventsList {
-    cursor: isize,
+    cursor: u32,
     events: Vec<String>,
 }
 
@@ -29,8 +29,8 @@ impl std::fmt::Display for EventsList {
 
 #[derive(Deserialize)]
 pub struct EventsQuery {
-    cursor: Option<isize>,
-    limit: Option<isize>,
+    cursor: Option<u32>,
+    limit: Option<u32>,
 }
 
 #[utoipa::path(
@@ -38,8 +38,8 @@ pub struct EventsQuery {
     path = EVENTS_ROUTE,
     tag = "Events",
     params(
-        ("cursor" = i64, Query, description = "Cursor"),
-        ("limit" = isize, Query, description = "Limit the number of results, (default 500, maximum 1000)")
+        ("cursor" = u32, Query, description = "Cursor"),
+        ("limit" = u32, Query, description = "Limit the number of results, (default 500, maximum 1000)")
     ),
     responses(
         (
@@ -70,7 +70,7 @@ pub async fn get_events_handler(Query(q): Query<EventsQuery>) -> Result<Response
     Ok(response)
 }
 
-fn parse_query(q: &EventsQuery) -> Result<(isize, Option<isize>), Error> {
+fn parse_query(q: &EventsQuery) -> Result<(u32, Option<u32>), Error> {
     let limit = q.limit.unwrap_or(500).min(1000);
     let cursor = q.cursor;
